# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "StellarSettle API",
    "version": "1.0.0",
    "description": "API specification for StellarSettle backend including Soroban contract interactions."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "BuyLimitResponse": {
        "type": "object",
        "properties": {
          "keyId": {
            "type": "string"
          },
          "maxBuyPerTx": {
            "type": "string",
            "description": "Maximum amount purchasable in a single buy transaction, in base units"
          },
          "maxBuyPerDay": {
            "type": "string",
            "description": "Maximum amount purchasable per wallet per day, in base units"
          },
          "currentSupply": {
            "type": "string",
            "description": "Current key supply, in base units"
          },
          "curveType": {
            "type": "string"
          },
          "configVersion": {
            "type": "integer"
          },
          "isActive": {
            "type": "boolean"
          },
          "contractAddress": {
            "type": "string"
          },
          "cachedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreatorKeyDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BuyLimitResponse"
          },
          {
            "type": "object",
            "properties": {
              "creatorId": {
                "type": "string"
              },
              "creatorWallet": {
                "type": "string"
              },
              "createdAt": {
                "type": "string",
                "format": "date-time"
              },
              "updatedAt": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        ]
      },
      "CurveMigration": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "migrationId": {
            "type": "string",
            "description": "On-chain proposal id"
          },
          "keyId": {
            "type": "string"
          },
          "proposalId": {
            "type": "string"
          },
          "contractAddress": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "executed",
              "expired"
            ]
          },
          "proposedParams": {
            "type": "object",
            "additionalProperties": true
          },
          "appliedParams": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          },
          "timelockExpiry": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "timelockExpired": {
            "type": "boolean"
          },
          "executable": {
            "type": "boolean"
          },
          "proposedAt": {
            "type": "string",
            "format": "date-time"
          },
          "executedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "proposalTxHash": {
            "type": "string",
            "nullable": true
          },
          "executionTxHash": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CurveMigrationList": {
        "type": "object",
        "properties": {
          "pending": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CurveMigration"
            }
          },
          "executed": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CurveMigration"
            }
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CurveMigration"
            }
          },
          "total": {
            "type": "integer"
          }
        }
      },
...
      "RoyaltyTransfer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "keyAddress": {
            "type": "string",
            "description": "Creator key contract that split the royalty."
          },
          "buyerWallet": {
            "type": "string",
            "nullable": true,
            "description": "Buyer that paid the transfer, when the event carried it."
          },
          "amount": {
            "type": "string",
            "pattern": "^\\d+(\\.\\d{1,4})?$",
            "description": "Royalty amount credited in token units."
          },
          "txHash": {
            "type": "string",
            "nullable": true,
            "description": "Transaction that paid the royalty."
          },
          "ledgerSequence": {
            "type": "string",
            "nullable": true,
            "description": "Ledger the royalty was paid in."
          },
          "paidAt": {
            "type": "string",
            "nullable": true,
            "description": "RFC 3339 payment time."
          }
        },
        "required": [
          "id",
          "keyAddress",
          "amount"
        ]
      },
      "RoyaltyClaim": {
        "type": "object",
        "properties": {
          "claimId": {
            "type": "string",
            "description": "On-chain claim identifier."
          },
          "keyAddress": {
            "type": "string",
            "nullable": true,
            "description": "Creator key the royalty was accrued under."
          },
          "amount": {
            "type": "string",
            "pattern": "^\\d+(\\.\\d{1,4})?$",
            "description": "Royalty amount claimed."
          },
          "txHash": {
            "type": "string",
            "nullable": true,
            "description": "Transaction that claimed the royalty."
          },
          "ledgerSequence": {
            "type": "string",
            "nullable": true,
            "description": "Ledger the claim settled in."
          },
          "claimedAt": {
            "type": "string",
            "nullable": true,
            "description": "RFC 3339 claim time."
          }
        },
        "required": [
          "claimId",
          "amount"
        ]
      },
      "RoyaltyEarnings": {
        "type": "object",
        "properties": {
          "creatorWallet": {
            "type": "string"
          },
          "totalEarned": {
            "type": "string",
            "pattern": "^\\d+(\\.\\d{1,4})?$",
            "description": "Lifetime royalty accruals."
          },
          "totalClaimed": {
            "type": "string",
            "pattern": "^\\d+(\\.\\d{1,4})?$",
            "description": "Lifetime royalties claimed."
          },
          "pending": {
            "type": "string",
            "pattern": "^\\d+(\\.\\d{1,4})?$",
            "description": "totalEarned - totalClaimed, floored at zero."
          },
          "transferCount": {
            "type": "integer"
          },
          "claimCount": {
            "type": "integer"
          },
          "transfers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoyaltyTransfer"
            }
          },
          "computedAt": {
            "type": "string",
            "description": "RFC 3339 computation time."
          }
        },
        "required": [
          "creatorWallet",
          "totalEarned",
          "totalClaimed",
          "pending",
          "transfers",
          "computedAt"
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/v1/keys/{id}": {
      "get": {
        "summary": "Get creator key detail",
        "description": "Returns the creator key configuration, including the per-transaction buy limit. Public read; the limit is enforced on-chain.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Creator key id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/CreatorKeyDetail"
                    }
                  }
                }
              }
            }
          },
          "Creator key not found": {
            "description": "Creator key not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "Bad Request": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/keys/{id}/buy-limit": {
      "get": {
        "summary": "Get per-transaction buy limit for a creator key",
        "description": "Pre-trade check: returns the per-transaction buy cap (max_buy_per_tx) and the current key supply so the frontend can clamp the buy input before submitting a transaction. Public read, cached for 60 seconds and invalidated when a key configuration update event is indexed.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Creator key id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/BuyLimitResponse"
                    }
                  }
                }
              }
            }
          },
          "Creator key not found": {
            "description": "Creator key not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "Bad Request": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/keys/{id}/curve-migrations": {
      "get": {
        "summary": "List bonding curve migrations for a creator key",
        "description": "Returns the migration proposals indexed from CurveMigrationProposed / CurveMigrationExecuted events, each with the proposed parameters, timelock expiry and status. Executed migrations are visible to any authenticated user; migrations still behind the timelock are creator-only.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Creator key id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "description": "Filter by status; `pending` requires creator access",
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "pending",
                "executed"
              ],
              "default": "all"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/CurveMigrationList"
                    }
                  }
                }
              }
            }
          },
          "Creator access required": {
            "description": "Creator access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "Authentication required": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "Bad Request": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/royalties/earnings": {
      "get": {
        "summary": "Creator royalty earnings",
        "description": "Lifetime royalty accruals, claims and the outstanding balance for the authenticated creator, with a per-transfer breakdown. Creator-only (seller or both) and scoped to the token's wallet. Figures are cached for 30 seconds and invalidated by royalty events.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "transfer_limit",
            "in": "query",
            "required": false,
            "description": "Maximum number of royalty transfers to return (1-200).",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/RoyaltyEarnings"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated wallet is not a creator",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/royalties/history": {
      "get": {
        "summary": "Creator royalty claim history",
        "description": "Royalties the authenticated creator has claimed, newest first, with the on-chain transaction that paid them. Cursor paginated and creator-only.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Page size (1-100).",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "description": "Opaque cursor from the previous page's meta.nextCursor.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "success": {
                    "type": "boolean"
                  },
                  "data": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/RoyaltyClaim"
                    }
                  },
                  "meta": {
                    "type": "object",
                    "properties": {
                      "hasMore": {
                        "type": "boolean"
                      },
                      "nextCursor": {
                        "type": "string",
                        "nullable": true,
                        "description": "Opaque cursor for the next page."
                      },
                      "limit": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The authenticated wallet is not a creator",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}